requirements.txt
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.1
Requests==2.31.0
transformers==4.37.2
